Drastically improve efficiency of tripole generation
In the tripole, we use the make_topo_gen routine. This was allocating
two arrays with 20,000,000 elements to hold topography for the
patch. The target grid is divided into blocks of 100x25 points, and
for each point in these blocks, for ever block, both of the large
arrays were being completely zeroed. This meant that performance was
almost completely dominated by memset.

Isolating to just 10 of these blocks (but including NetCDF input and
output), I saw 98.2% of the CPU time spent in memset, taking over 2
minutes to process. With this patch, it takes about 5 seconds to
process the blocks, where most of that time is spent in KD tree
generation and NetCDF input/output. An entire invocation of gen_topo
from GEBCO to a 1/10 full-globe grid went down from around 2 hours to
6 minutes.